backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Google Drive API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Redis for caching and task queue
redis==5.0.1
celery==5.3.4

# Scheduling
APScheduler==3.10.4
watchdog==3.0.0

# File handling
aiofiles==23.2.1
tqdm==4.66.1

# QR Code generation
qrcode[pil]==7.4.2

# CORS and Security
python-cors==1.0.0

# Utils
pydantic==2.5.2
httpx==0.25.2
pyyaml==6.0.1
click==8.1.7